pizzeria_app/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
python-dotenv==1.1.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
